Consolidate UI from 13 pages to 6 for improved navigation
Major UI/UX improvements:
- Consolidate 13 pages into 6 focused pages
- Promote Fleet & Stacks as primary view (page 1) with Stacks as default tab
- Merge container-related pages into unified Containers page with tabs:
  - All Containers, Health & Alerts, Images, Networks, Container Details
- Merge logging pages into unified Logs page with tabs:
  - Live Logs (Docker), Search Logs (Kibana)
- Merge AI features into AI Operations page with tabs:
  - Insights Dashboard, Self-Healing, Metrics, Anomalies
- Update Home page with quick navigation cards and cleaner layout
- Move Trace Explorer into Settings as Developer Tools tab
- Reduce code duplication and improve maintainability

New page structure:
1. Fleet & Stacks - Edge agents and stack deployments
2. Containers - Health, images, networks, details
3. Logs - Live and searchable logs
4. AI Operations - Insights, self-healing, metrics
5. LLM Assistant - Chat interface
6. Settings - Config, backups, tracing, about
